requirements.txt
streamlit
opencv-python
numpy
pandas
matplotlib
Pillow
scipy
PyMuPDF
fpdf
google-generativeai
